pyproject.toml
[project]
name = "lucid-finance-webapp"
version = "0.1.0"
description = "Personal budgeting application with ETL pipeline for bank transactions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "cryptography>=46.0.4",
    "fastapi>=0.128.3",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt==4.1.3",
    "openpyxl>=3.1.0",
]

[project.scripts]
lucid-etl = "backend.data_pipeline.pipeline:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
